ci: add no circular dependency rule to tslint
And fix the linting issues.
