+++ packages/angular_devkit/schematics/src/engine/index.ts
@@ -5,5 +5,6 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-
-export * from './architect';
+export * from './engine';
+export * from './interface';
+export * from './schematic';
