+++ packages/angular_devkit/schematics/src/engine/interface.ts
@@ -10,8 +10,36 @@ import { Observable } from 'rxjs';
 import { Url } from 'url';
 import { FileEntry, MergeStrategy, Tree } from '../tree/interface';
 import { Workflow } from '../workflow';
-import { TaskConfigurationGenerator, TaskExecutor, TaskId } from './task';
 
+
+export interface TaskConfiguration<T = {}> {
+  name: string;
+  dependencies?: Array<TaskId>;
+  options?: T;
+}
+
+export interface TaskConfigurationGenerator<T = {}> {
+  toConfiguration(): TaskConfiguration<T>;
+}
+
+export type TaskExecutor<T = {}>
+  = (options: T | undefined, context: SchematicContext) => Promise<void> | Observable<void>;
+
+export interface TaskExecutorFactory<T> {
+  readonly name: string;
+  create(options?: T): Promise<TaskExecutor> | Observable<TaskExecutor>;
+}
+
+export interface TaskId {
+  readonly id: number;
+}
+
+export interface TaskInfo {
+  readonly id: number;
+  readonly priority:  number;
+  readonly configuration: TaskConfiguration;
+  readonly context: SchematicContext;
+}
 
 /**
  * The description (metadata) of a collection. This type contains every information the engine
