+++ packages/angular_devkit/schematics/src/index.ts
@@ -13,16 +13,7 @@ import { branch, empty, merge, optimize, partition } from './tree/static';
 export { SchematicsException } from './exception/exception';
 
 export * from './tree/action';
-export * from './engine/collection';
-export * from './engine/engine';
-export * from './engine/interface';
-export {
-  TaskConfiguration,
-  TaskConfigurationGenerator,
-  TaskExecutor,
-  TaskExecutorFactory,
-  TaskId,
-} from './engine/task';
+export * from './engine';
 export * from './exception/exception';
 export * from './tree/interface';
 export * from './rules/base';
