+++ packages/angular_devkit/schematics/tools/node-modules-test-engine-host.ts
@@ -5,7 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import { TaskConfiguration, TaskConfigurationGenerator, TaskId } from '../src/engine/task';
+import { TaskConfiguration, TaskConfigurationGenerator, TaskId } from '../src/engine';
 import { FileSystemSchematicContext } from './description';
 import { NodeModulesEngineHost } from './node-module-engine-host';
